I suggest having these changes in the ControlBase class.
The first one, which is about disabling the control readiness check, will allow to increase the tests' execution speed, and the second one just includes an ability to wait for a control to exist in the UI (because WaitForControlReady is definitely not enough).

All meaningful changes are in the ControlBase class. Other classes are affected by renaming the WaitForControlReady method to WaitForControlReadyIfNecessary (it is also made protected).